Basic Error Reporting for Queue Size
When running the simulation something must be done to enforce the fact
that this value is wrong. As there is no minimum queue size I recommend
that a notification is posed to the user 'your value for Queue Size
exceeds the maximum, the maximum value of X will be used instead' and
then update the JTextField to represent that.